Add vllm logs to nop harness results, move txt file to analysis dir (#164)
